fix(example): %u format mismatch on recovered_count uint32_t
Same -Werror=format= issue as the lib fix in dd995ab — the full-demo
example's MQTT_CONNECTED handler logged recovered_count with %u, but on
Xtensa GCC uint32_t == long unsigned int. Cast to unsigned long + use
%lu so the format width agrees across hosts.

Hello-world example builds clean now; this should make full-demo green.

+++ examples/esp-idf-full-demo/main/main.c
@@ -65,7 +65,10 @@ static void on_scd_event(scadable_event_t e, void *user) {
     (void)user;
     switch (e.type) {
     case SCADABLE_EVT_CONNECTED:
-        SCADABLE_LOG_INFO("MQTT connected (recovered_count=%u)", e.connected.recovered_count);
+        // recovered_count is uint32_t (long unsigned int on Xtensa) — cast to
+        // unsigned long so the %lu format width matches across hosts.
+        SCADABLE_LOG_INFO("MQTT connected (recovered_count=%lu)",
+                          (unsigned long)e.connected.recovered_count);
         break;
     case SCADABLE_EVT_DISCONNECTED:
         SCADABLE_LOG_WARN("MQTT disconnected");
